pdf_modifier/requirements.txt
python-docx
openai